+++ MHD/C_src/KEMO_GL/draw_fieldlines.c
@@ -15,7 +15,7 @@ void const_fieldlines_buffer(struct psf_data *fline_s, struct fline_menu_val *fl
     set_color_code_for_fieldlines(fline_s, fline_m);
 
     if(fline_m->fieldline_type == IFLAG_PIPE){
-        int num_patch = count_fieldtubes_to_buf(ncorner, fline_s);
+        long num_patch = count_fieldtubes_to_buf(ncorner, fline_s);
         set_buffer_address_4_patch(ITHREE*num_patch, FLINE_tube_buf);
         if(FLINE_tube_buf->num_nod_buf> 0){
             resize_strided_buffer(FLINE_tube_buf);
@@ -23,7 +23,7 @@ void const_fieldlines_buffer(struct psf_data *fline_s, struct fline_menu_val *fl
         };
     };
 
-    int num_edge = count_fieldlines_to_buf(fline_s);
+    long num_edge = count_fieldlines_to_buf(fline_s);
     set_buffer_address_4_patch(ITWO*num_edge, FLINE_line_buf);
     if(FLINE_line_buf->num_nod_buf>0){
         resize_strided_buffer(FLINE_line_buf);
