+++ MHD/C_src/KEMO_GL/set_mesh_node_2_gl_buf.h
@@ -13,7 +13,7 @@
 
 /* prototypes */
 int count_mesh_node_to_buf(struct viewer_mesh *mesh_s, struct mesh_menu_val *mesh_m);
-int set_mesh_node_to_buf(struct viewer_mesh *mesh_s, struct mesh_menu_val *mesh_m,
-			struct gl_strided_buffer *mesh_buf);
+long set_mesh_node_to_buf(struct viewer_mesh *mesh_s, struct mesh_menu_val *mesh_m,
+                          struct gl_strided_buffer *mesh_buf);
 
 #endif
