+++ MHD/Unused/kemoviewer_GLUI/kemo_mesh_viewer_glui.cpp
@@ -752,7 +752,7 @@ static void make_3rd_level_psf_menu(){
 	int num_psf =     kemoview_get_PSF_num_loaded();
 	int num_fld =     kemoview_get_PSF_num_field();
 	int if_psf =      kemoview_get_PSF_field_id();
-	int num_comp =    kemoview_get_PSF_num_component(kemo_sgl, if_psf);
+	int num_comp =    (int) kemoview_get_PSF_num_component(kemo_sgl, if_psf);
 	int iflag_solid = kemoview_get_PSF_draw_flags(kemo_sgl, PSFSOLID_TOGGLE);
 	int iflag_grid =  kemoview_get_PSF_draw_flags(kemo_sgl, PSFGRID_TOGGLE);
 	
@@ -825,7 +825,7 @@ static void make_2nd_level_psf_menu(){
 	int num_fld =     kemoview_get_PSF_num_field();
 	int if_psf =      kemoview_get_PSF_field_id();
 	int ic_psf =      kemoview_get_PSF_component_id();
-	int num_comp =    kemoview_get_PSF_num_component(kemo_sgl, if_psf);
+	int num_comp =    (int) kemoview_get_PSF_num_component(kemo_sgl, if_psf);
 	int iflag_solid = kemoview_get_PSF_draw_flags(kemo_sgl, PSFSOLID_TOGGLE);
 	int iflag_grid =  kemoview_get_PSF_draw_flags(kemo_sgl, PSFGRID_TOGGLE);
